Add BaseWindshaft source for Dataset and SQL
