+++ src/api/source/dataset.js
@@ -1,9 +1,9 @@
 import * as util from '../util';
-import Base from './base';
+import BaseWindshaft from './base-windshaft';
 import CartoValidationError from '../error-handling/carto-validation-error';
 
 
-export default class Dataset extends Base {
+export default class Dataset extends BaseWindshaft {
 
     /**
      * Create a carto.source.Dataset.
